backend/requirements.txt
Flask
flask-cors
groq
python-dotenv
requests
jinja2
